requirements.txt
flask
twilio
python-dotenv
langchain
langchain-huggingface
langchain-pinecone
pinecone
openai